+++ enterprise/static/enterprise/js/grant_data_sharing_permissions.js
@@ -56,7 +56,8 @@ function loadConsentPage() {
         successUrl: successUrl,
         failureUrl: failureUrl,
         courseId: courseId,
-        programId: programId
+        programId: programId,
+        licenseUuid: licenseUuid,
     };
     analytics.track("edx.bi.user.consent_form.shown", formDetails);
     analytics.trackForm($("#data-sharing"), function () {